requirements.txt
gunicorn==20.1.0
djangorestframework==3.14.0
